Updated settings UI
