+++ Tempo/Views/PlayerCellView.swift
@@ -119,8 +119,8 @@ class PlayerCellView: UIView {
 					let appDelegate = UIApplication.shared.delegate as! AppDelegate
 					let playerNav = appDelegate.navigationController
 					let revealVC = appDelegate.revealVC
-					let spotifyVC = appDelegate.spotifyVC
-					playerNav.setViewControllers([spotifyVC], animated: true)
+					let settingsVC = appDelegate.settingsVC
+					playerNav.setViewControllers([settingsVC], animated: true)
 					revealVC.setFrontViewPosition(.left, animated: true)
 				}
 			}
